requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
airflowctl==0.2.9
alembic==1.13.1
amqp==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
apache-airflow==2.9.2
apache-airflow-providers-apache-spark==4.8.1
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.0
apache-airflow-providers-fab==1.1.1
apache-airflow-providers-ftp==3.9.1
apache-airflow-providers-http==4.11.1
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-postgres==5.11.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
appdirs==1.4.4
argcomplete==3.4.0
asgiref==3.8.1
async-timeout==4.0.3
asyncssh==2.14.2
atpublic==4.1.0
attrs==23.2.0
Babel==2.15.0
billiard==4.2.0
blinker==1.8.2
cachelib==0.9.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
configobj==5.0.8
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==42.0.8
Deprecated==1.2.14
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docutils==0.21.2
dpath==2.2.0
dulwich==0.22.1
dvc==3.51.2
dvc-data==3.15.1
dvc-hdfs==3.0.0
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.20.0
dvc-task==0.4.0
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.2.1
filelock==3.15.1
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.4.1
fsspec==2024.6.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
google-re2==1.1.20240601
googleapis-common-protos==1.63.1
grandalf==0.8
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
gto==1.7.1
gunicorn==22.0.0
h11==0.14.0
hdfs3==0.3.1
httpcore==0.17.3
httpx==0.24.1
hydra-core==1.3.2
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflection==0.5.1
iterative-telemetry==0.0.8
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
lazy-object-proxy==1.10.0
limits==3.12.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.3.0
multidict==6.0.5
networkx==3.3
numpy==2.0.0
omegaconf==2.3.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
orjson==3.10.5
packaging==23.2
pathspec==0.12.1
pendulum==3.0.0
platformdirs==3.11.0
pluggy==1.5.0
prison==0.2.1
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyarrow==16.1.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydot==2.0.0
pygit2==1.15.0
Pygments==2.18.0
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.1.2
pyspark==3.5.1
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scmrepo==3.3.5
semver==3.0.2
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
sqltrie==0.11.0
tabulate==0.9.0
tenacity==8.4.1
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.1
tomlkit==0.12.5
tqdm==4.66.4
typer==0.9.4
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.2.2
vine==5.1.0
voluptuous==0.14.2
wcwidth==0.2.13
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.19.2
